% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mich_sets.R
\name{cred_set}
\alias{cred_set}
\title{Single Change-Point Posterior Credible Set}
\usage{
cred_set(prob, level)
}
\arguments{
\item{prob}{A numeric vector. A vector of posterior probabilities for the
location of the change-point.}

\item{level}{A scalar. A single number in (0,1) that gives the lower bound
for the probability that the credible set contains a change-point.}
}
\value{
A vector. A Level \code{level} posterior credible set for the location of
a single change-point.
}
\description{
The function \code{cred_set()} takes a length T vector of posterior change-point
location probabilities \code{prob} and a coverage level \code{level}, and returns the
smallest set of indices \code{s} such that \code{prob[s] > level}.
}
