% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_steps}
\alias{daily_steps}
\title{Davis Berlind's Daily Step Count}
\format{
\subsection{\code{daily_steps}}{

A data frame with 3,247 rows and 2 columns:
\describe{
\item{day}{Date}
\item{total_steps}{Total steps in day}
}
}
}
\source{
My Apple Health app.
}
\usage{
daily_steps
}
\description{
Davis Berlind's total daily steps between 9-12-2016 and 08-15-2025
}
\keyword{datasets}
