% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{integer_check}
\alias{integer_check}
\title{Integer value check.}
\usage{
integer_check(n)
}
\arguments{
\item{n}{An integer.}
}
\value{
An integer. If no error is thrown during evaluation, then
\code{integer_check()} returns the \code{n} argument.
}
\description{
\code{integer_check()} throws an error if \code{n} is not equal to a positive valued
integer.
}
