% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ion_channel}
\alias{ion_channel}
\title{Ion Channel Voltage}
\format{
\subsection{\code{ion_channel}}{

A data frame with 2,956 rows and 1 columns:
\describe{
\item{y}{Ionic current recordings}
}
}
}
\source{
\href{https://www.uni-goettingen.de/en/213067.html}{Steinem lab (Institute of Organic and Biomolecular Chemistry, University of Gottingen)}
}
\usage{
ion_channel
}
\description{
Ionic current recordings from a porin in the outer membrane of a bacterial
cell using the voltage clamp technique. Data have been processed so that they
include every 11th observation.
}
\keyword{datasets}
