% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_bar_fn}
\alias{lambda_bar_fn}
\title{Expected Precision Signal}
\usage{
lambda_bar_fn(u, v, prob)
}
\arguments{
\item{u}{A numeric vector. Length \eqn{T} vector of conditional shape
parameters}

\item{v}{A numeric vector. Length \eqn{T} vector of conditional rate
parameters}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric vector. A length \eqn{T} vector of \eqn{E[\lambda_t]}.
}
\description{
Given that \eqn{V(y_t|s,\tau)=1/\lambda_t=1/s^{I(t\geq \tau)}},
\eqn{P(\tau=t)=\pi_t}, and \eqn{\bar{u}_t/\bar{v}_t=E[s|\tau=t]},
\code{lambda_bar_fn()} calculates \eqn{E[\lambda_t]} as
\eqn{1 - \Sigma_{t'=1}^t \pi_{t'}(1 - \bar{v}_{t'} / \bar{u}_{t'})}.
}
\keyword{internal}
