% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logical_check}
\alias{logical_check}
\title{Logical value check.}
\usage{
logical_check(x)
}
\arguments{
\item{x}{A logical.}
}
\value{
A logical. If no error is thrown during evaluation, then
\code{logical_check()} returns the \code{x} argument.
}
\description{
\code{logical_check()} throws an error if \code{x} is not equal to either \code{TRUE} or
\code{FALSE}.
}
