% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_scp}
\alias{mean_scp}
\title{Mean Single Change-Point Model}
\usage{
mean_scp(y, lambda, omega, log_pi)
}
\arguments{
\item{y}{A numeric vector. \eqn{T} observations with a single variance
change.}

\item{lambda}{A numeric vector. Precision vector of \code{y}.}

\item{omega}{A positive scalar. Prior precision parameter.}

\item{log_pi}{A numeric vector. Vector of log prior probabilities for the
location of the change-point.}
}
\value{
A list. A list of posterior parameters including the mean \code{b_bar},
precision \code{omega_bar}, and posterior probabilities of the change-point
location \code{pi_bar}.
}
\description{
Implementation of the univariate Mean-SCP model from Berlind, Cappello, and
Madrid Padilla (2025). The function \code{mean_scp()} takes a length \eqn{T}
vector \eqn{y_{1:T}} with a single mean change and returns the posterior
distribution of the change-point.
}
