% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanvar_scp}
\alias{meanvar_scp}
\title{Mean-Variance Single Change-Point Model}
\usage{
meanvar_scp(y, lambda, omega, u_bar, lgamma_u_bar, v, log_pi)
}
\arguments{
\item{y}{A numeric vector. \eqn{T} observations with a single joint mean and
variance change.}

\item{lambda}{A numeric vector. Known trend component of precision of \code{y}.}

\item{omega}{A positive scalar. Prior precision parameter.}

\item{u_bar}{A numeric vector. Posterior shape parameters equal to
\eqn{u_0 + T - t + 1} for each \eqn{t}.}

\item{lgamma_u_bar}{A numeric vector. Log gamma function evaluated at u_bar.}

\item{v}{A numeric vector. Vector of prior rate parameters for each \eqn{t}.}

\item{log_pi}{A numeric vector. Vector of log prior probabilities for the
location of the change-point.}
}
\value{
A list. A list of posterior parameters including the  mean \code{b_bar},
precision \code{omega_bar}, rate \code{v_bar}, and posterior probabilities of the
change-point location \code{pi_bar}.
}
\description{
Implementation of the MeanVar-SCP model from Berlind, Cappello, and Madrid
Padilla (2025). The function \code{meanvar_scp()} takes a length \eqn{T} vector
\eqn{y_{1:T}} with a single joint mean and variance change and returns the
posterior distribution of the change-point.
}
