% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mich_cpp}
\alias{mich_cpp}
\title{MICH Algorithm}
\usage{
mich_cpp(
  y,
  J,
  L,
  K,
  mu_0,
  lambda_0,
  fit_intercept,
  fit_scale,
  refit,
  max_iter,
  verbose,
  tol,
  omega_j,
  u_j,
  v_j,
  log_pi_j,
  pi_bar_j,
  log_pi_bar_j,
  b_bar_j,
  omega_bar_j,
  u_bar_j,
  v_bar_j,
  lgamma_u_bar_j,
  digamma_u_bar_j,
  omega_l,
  log_pi_l,
  pi_bar_l,
  log_pi_bar_l,
  b_bar_l,
  omega_bar_l,
  u_k,
  v_k,
  log_pi_k,
  pi_bar_k,
  log_pi_bar_k,
  u_bar_k,
  v_bar_k,
  lgamma_u_bar_k,
  digamma_u_bar_k
)
}
\arguments{
\item{y}{A numeric vector. Length \eqn{T} vector of observations.}

\item{L, K, J}{Integers. The number of mean, variance, and mean-variance
change-points to include in the model.}

\item{mu_0}{A scalar. Intercept parameter initialization.}

\item{lambda_0}{A scalar. Baseline scale parameter initialization.}

\item{fit_intercept}{A logical. If \code{fit_intercept == TRUE}, then an
intercept is estimated and \code{mu_0} gets updated.}

\item{fit_scale}{A logical. If \code{fit_scale == TRUE}, then an initial scale is
estimated and \code{lambda_0} gets updated.}

\item{refit}{A logical. If \code{refit == TRUE}, then the MICH algorithm is
initialized using the provided posterior parameters, otherwise the null
model \eqn{\mu_t = 0}, \eqn{\lambda_t = 1} and
\eqn{\bar{\pi}_{\ell t} = \bar{\pi}_{k t} = \bar{\pi}_{j t}  = 1/T} is
used as the initialization.}

\item{max_iter}{An integer. Maximum number of iterations. If ELBO does not
converge before \code{max_iter} is reached, then \code{converged == FALSE} in the
returned fit object.}

\item{verbose}{A logical. If \code{verbose == TRUE}, then value of the ELBO is
printed every 5000th iteration.}

\item{tol}{A scalar. Convergence tolerance for relative increase in ELBO.}

\item{omega_j, u_j, v_j}{Scalars. Prior precision, shape, and rate parameters
for meanvar-scp components of model.}

\item{log_pi_j}{A numeric matrix. A \eqn{T \times J} matrix of prior log
change-point location probabilities for each of the \eqn{J} mean-variance
change-points.}

\item{pi_bar_j, log_pi_bar_j}{Numeric matrices. \eqn{T \times J} matrices of
initialized posterior change-point location probabilities and their log
evaluations for the \eqn{J} mean-variance change-points.}

\item{b_bar_j}{A numeric matrix. A \eqn{T \times J} matrix of initialized
posterior mean parameters of the \eqn{J} mean-variance change-points.}

\item{omega_bar_j}{A numeric matrix. A \eqn{T\times J} matrix of initialized
posterior precision parameters of the \eqn{J} mean-variance change-points.}

\item{u_bar_j, lgamma_u_bar_j, digamma_u_bar_j}{Numeric vectors. Length
\eqn{T} vectors of posterior shape parameters for the meanvar-scp model
components and their log-gamma and digamma evaluations.}

\item{v_bar_j}{A numeric matrix. A \eqn{T \times J} matrix of initialized
posterior rate parameters for the \eqn{J} mean-variance change-points.}

\item{omega_l}{A scalar. Prior precision parameter for mean-scp components
of model.}

\item{log_pi_l}{A numeric matrix. A \eqn{T \times L} matrix of prior log
change-point location probabilities for each of the \eqn{L} mean
change-points.}

\item{pi_bar_l, log_pi_bar_l}{Numeric matrices. \eqn{T \times L} matrices of
initialized posterior change-point location probabilities and their log
evaluations for the \eqn{L} mean change-points.}

\item{b_bar_l}{A numeric matrix. A \eqn{T \times L} matrix of initialized
posterior mean parameters of the \eqn{L} mean change-points.}

\item{omega_bar_l}{A numeric matrix. A \eqn{T\times L} matrix of initialized
posterior precision parameters of the \eqn{L} mean change-points.}

\item{u_k, v_k}{Scalars. Prior shape and rate parameters for var-scp
components of model.}

\item{log_pi_k}{A numeric matrix. A \eqn{T \times K} matrix of prior log
change-point location probabilities for each of the \eqn{K} variance
change-points.}

\item{pi_bar_k, log_pi_bar_k}{Numeric matrices. \eqn{T \times K} matrices of
initialized posterior change-point location probabilities and their log
evaluations for the \eqn{K} variance change-points.}

\item{u_bar_k, lgamma_u_bar_k, digamma_u_bar_k}{Numeric vectors. Length
\eqn{T} vectors of posterior shape parameters for the var-scp model
components and their log-gamma and digamma evaluations.}

\item{v_bar_k}{A numeric matrix. A \eqn{T \times K} matrix of initialized
posterior rate parameters for the \eqn{K} variance change-points.}
}
\value{
A List. Parameters of the variational approximation the MICH
posterior distribution, including:
\itemize{
\item \code{y}: A numeric vector. Original data set.
\item \code{residual}: A numeric vector. Residual \eqn{\mathbf{r}_{1:T}} after
subtracting out each \eqn{E[\mu_{\ell t}]} and \eqn{E[\lambda_{j t} \mu_{j t}]/E[\lambda_{j t}]}
from \eqn{\mathbf{y}_{1:T}}.
\item \code{mu}: A numeric vector. Posterior estimate of
\eqn{\Sigma_{\ell=1}^L E[\mu_{\ell,1:T}|\mathbf{y}_{1:T}] + \Sigma_{j=1}^J E[\mu_{j,1:T}|\mathbf{y}_{1:T}]}.
\item \code{lambda}: A numeric vector. Posterior estimate of
\eqn{\Pi_{\\k=1}^K E[\lambda_{k,1:T}|\mathbf{y}_{1:T}] \times \Pi_{j=1}^J E[\lambda_{j,1:T}|\mathbf{y}_{1:T}]}.
\item \code{delta}: A numeric vector. Posterior variance correction term (see
(B.6) or Berlind, Cappello, and Madrid Padilla (2025)).
\item \code{converged}: A logical. Indicates whether relative increase in the ELBO
is less than \code{tol}.
\item \code{elbo}: A numeric vector. Value of the ELBO after each iteration.
\item \code{mu_0}: A scalar. Estimate of the intercept.
\item \code{lambda_0}: A scalar. Estimate of the initial precision.
\item \code{J}, \code{L}, \code{K}: Integers. number of mean, variance, and mean-variance
components.
\item \code{J_model}: A list. A list of the posterior parameters for each of the
\eqn{J} meanvar-scp components (only included in \code{J > 0}).
\item \code{L_model}: A list. A list of the posterior parameters for each of the
\eqn{L} mean-scp components (only included in \code{L > 0}).
\item \code{K_model}: A list. A list of the posterior parameters for each of the
\eqn{K} var-scp components (only included in \code{K > 0}).
}
}
\description{
Implementation of Algorithms 1 & 2 in Berlind, Cappello, and Madrid Padilla
(2025). This algorithm takes a sequence of \eqn{T} observations
\eqn{\mathbf{y}_{1:T}}, and iteratively fits \eqn{L} mean-scp models,
\eqn{K} var-scp models, and \eqn{J} meanvar-scp models resulting in a
variational approximation to the posterior distribution of
the \eqn{L} mean, \eqn{K} variance, and \eqn{J} mean and variance
change-points. The algorithm terminates once the percentage increase in the
ELBO falls below \code{tol}.
}
