% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mich_vector.R
\name{mich_vector}
\alias{mich_vector}
\title{Univariate Multiple Independent Change-Point (MICH) Model}
\usage{
mich_vector(
  y,
  fit_intercept,
  fit_scale,
  standardize,
  J,
  L,
  K,
  J_auto,
  L_auto,
  K_auto,
  J_max,
  L_max,
  K_max,
  pi_j_weighted,
  pi_l_weighted,
  pi_k_weighted,
  tol,
  verbose,
  max_iter,
  reverse,
  detect,
  merge_level,
  merge_prob,
  restart,
  n_search,
  increment,
  omega_j,
  u_j,
  v_j,
  log_pi_j,
  omega_l,
  log_pi_l,
  u_k,
  v_k,
  log_pi_k
)
}
\arguments{
\item{y}{A numeric vector. Length \eqn{T} vector of observations.}

\item{fit_intercept}{A logical. If \code{fit_intercept == TRUE}, then an
intercept is estimated, otherwise it is assumed  that
\eqn{\mu_0 = 0}.}

\item{fit_scale}{A logical. If \code{fit_scale == TRUE}, then the initial
precision is estimated, otherwise it is assumed  that \eqn{\lambda_0 = 1}.}

\item{standardize}{A logical. If \code{standardize == TRUE}, then \code{y} is centered
and rescaled before fitting.}

\item{L, K, J}{Integers. Respective number of mean-scp, var-scp, and
meanvar-scp components included in the model. If \code{L_auto == TRUE},
\code{K_auto == TRUE}, or \code{J_auto == TRUE} then \code{L}, \code{K}, and \code{J} lower bound
the number of each kind of change-point in the model.}

\item{L_auto, K_auto, J_auto}{Logicals. If \code{L_auto == TRUE},
\code{K_auto == TRUE}, and/or \code{J_auto == TRUE}, then \code{mich_vector()} returns the
\eqn{L} between \code{L} and \code{L_max}, the \eqn{K} between \code{K} and \code{K_max},
and/or the \eqn{J} between \code{J} and \code{J_max} that maximize the ELBO (see
Appendix C.4 of Berlind, Cappello, Madrid Padilla (2025)).}

\item{L_max, K_max, J_max}{Integers. If \code{L_auto == TRUE},
\code{K_auto == TRUE}, or \code{J_auto == TRUE} then \code{L_max}, \code{K_max}, and \code{J_max}
upper bound the number of each kind of change-point in the model.}

\item{pi_l_weighted, pi_k_weighted, pi_j_weighted}{Logicals. If \code{TRUE}, then
the weighted priors specified in Appendix C.2 of Berlind, Cappello, Madrid
Padilla (2025) are used.}

\item{tol}{A scalar. Convergence tolerance for relative increase in ELBO.}

\item{verbose}{A logical. If \code{verbose == TRUE} and \code{L_auto == FALSE},
\code{K_auto == FALSE}, and \code{J_auto == FALSE} then the value of the ELBO is
printed every 5000th iteration. If \code{verbose == TRUE} and any of \code{L_auto},
\code{K_auto}, or \code{J_auto} are \code{TRUE}, then the value of the ELBO is printed for
each combinatin of \eqn{(L,K,J)} as \code{mich_vector()} searches for the
parameterization that maximized the EBLO.}

\item{max_iter}{An integer. Maximum number of iterations. If ELBO does not
converge before \code{max_iter} is reached, then \code{converged == FALSE} in the
returned fit object.}

\item{reverse}{A logical. If \code{reverse == TRUE} then MICH is fit to
\eqn{\mathbf{y}_{T:1}} and the model parameters are reversed in
post-processing.}

\item{detect}{A scalar. The detection criteria. The \eqn{i^{\text{th}}}
component of the model detects a change-point only if the posterior
credible set for that component contains fewer than \code{detect} indices.}

\item{merge_level}{A scalar. A value between (0,1) for the significance level
to construct credible sets at when merging. A model component is only
considered to be a candidate for merging if its \code{merge_level}-level
credible set contains fewer than \code{detect} indices.}

\item{merge_prob}{A scalar. A value between (0,1) indicating the merge
criterion. If the posterior probability that two components identify the
same change is greater than \code{merge_level}, then those components are merged
(see Appendix C.3 of Berlind, Cappello, Madrid Padilla (2025)).}

\item{restart}{A logical. If \code{restart == TRUE} and \code{L_auto}, \code{K_auto}, or
\code{J_auto} are \code{TRUE}, then after \code{n_search} increments of \code{L}, \code{K}, and/or
\code{J}, if the ELBO has not increased, \code{mich_vector()} will restart by setting
the \code{L}, \code{K}, and \code{J} components to the null model initialization (except
for the components with maximum posterior probabilities > 0.9) then refit
and begin the search again.}

\item{n_search}{An integer. Grid search parameter. Number of times to
increment \code{L}, \code{K}, and/or \code{J} before terminating automatic procedure when
\code{L_auto}, \code{K_auto}, or \code{J_auto} are \code{TRUE}.}

\item{increment}{An integer. Number of components to increment \code{L}, \code{K} and
\code{J} by when \code{L_auto}, \code{K_auto}, or \code{J_auto} are \code{TRUE}.}

\item{omega_j}{A scalar. Prior precision parameter for meanvar-scp components
of the model.}

\item{u_j}{A scalar. Prior shape parameter for meanvar-scp components
of the model.}

\item{v_j}{A scalar. Prior rate parameter for meanvar-scp components
of the model.}

\item{log_pi_j}{A numeric matrix. A \eqn{T \times J} matrix of prior log
change-point location probabilities for each of the \eqn{J} mean-variance
change-points.}

\item{omega_l}{A scalar. Prior precision parameter for mean-scp components of
the model.}

\item{log_pi_l}{A numeric matrix. A \eqn{T \times L} matrix of prior log
change-point location probabilities for each of the \eqn{L} mean
change-points.}

\item{u_k}{A scalar. Prior shape parameter for var-scp components
of the model.}

\item{v_k}{A scalar. Prior rate parameter for var-scp components
of the model.}

\item{log_pi_k}{A numeric matrix. A \eqn{T \times K} matrix of prior log
change-point location probabilities for each of the \eqn{K} variance
change-points.}
}
\value{
A list. Parameters of the variational approximation the MICH
posterior distribution, including:
\itemize{
\item \code{y}: A numeric vector. Original data.
\item \code{L},\code{K},\code{J}: Integers. Number of mean-scp, var-scp, and meanvar-scp
components included in the model.
\item \code{residual}: A numeric vector. Residual \eqn{\tilde{\mathbf{r}}_{1:T}}
(see (B.4) of Berlind, Cappello, Madrid Padilla (2025)).
\item \code{mu}: A numeric vector. Posterior estimate of
\eqn{\Sigma_{\ell=1}^L E[\mu_{\ell,1:T}|\mathbf{y}_{1:T}] + \Sigma_{j=1}^J E[\mu_{j,1:T}|\mathbf{y}_{1:T}]}.
\item \code{lambda}: A numeric vector. Posterior estimate of
\eqn{\Pi_{k=1}^K E[\lambda_{k,1:T}|\mathbf{y}_{1:T}] \times \Pi_{j=1}^J E[\lambda_{j,1:T}|\mathbf{y}_{1:T}]}.
\item \code{delta}: A numeric vector. Posterior estimate of equation (B.4) of
Berlind, Cappello, Madrid Padilla (2025).
\item \code{mu_0}: A scalar. Estimate of the intercept.
\item \code{lambda_0}: A scalar. Estimate of the initial precision.
\item \code{elbo}: A numeric vector. Value of the ELBO after each iteration.
\item \code{converged}: A logical. Indicates whether relative increase in the ELBO
is less than \code{tol}.
\item \code{meanvar_model}: A list. List of meanvar-scp posterior parameters:
\itemize{
\item \code{pi_bar}: A numeric matrix. A \eqn{T \times J} matrix of posterior
change-point location probabilities.
\item \code{b_bar}: A numeric matrix. A \eqn{T \times J} matrix of posterior
mean parameters.
\item \code{omega_bar}: A numeric matrix. A \eqn{T \times J} matrix of posterior
precision parameters.
\item \code{v_bar}: A numeric matrix. A \eqn{T \times J} matrix of posterior
rate parameters.
\item \code{u_bar}: A numeric vector. A length \eqn{T} vector of posterior shape
parameters.
\item \code{mu_lambda_bar}: A numeric matrix. A \eqn{T \times J} matrix of scaled
posterior mean signals \eqn{E[\lambda_{jt}\mu_{jt}|\mathbf{y}_{1:T}]}.
\item \code{mu2_lambda_bar}: A numeric matrix. A \eqn{T \times J} matrix of scaled
posterior squared mean signals
\eqn{E[\lambda_{jt}\mu^2_{jt}|\mathbf{y}_{1:T}]}.
\item \code{lambda_bar}: A numeric matrix. A \eqn{T \times J} matrix of
posterior precision signals \eqn{E[\lambda_{jt}|\mathbf{y}_{1:T}]}.
}
\item \code{mean_model}: A list. List of mean-scp posterior parameters:
\itemize{
\item \code{pi_bar}: A numeric matrix. A \eqn{T \times L} matrix of posterior
change-point location probabilities.
\item \code{b_bar}: A numeric matrix. A \eqn{T \times L} matrix of posterior
mean parameters.
\item \code{omega_bar}: A numeric matrix. A \eqn{T \times L} matrix of posterior
precision parameters.
\item \code{mu_bar}: A numeric matrix. A \eqn{T \times L} matrix of posterior mean
signals \eqn{E[\mu_{\ell t}|\mathbf{y}_{1:T}]}.
\item \code{mu2_bar}: A numeric matrix. A \eqn{T \times L} matrix of posterior
squared mean signals \eqn{E[\mu_{\ell t}^2|\mathbf{y}_{1:T}]}.
}
\item \code{var_model}: A list. List of var-scp posterior parameters:
\itemize{
\item \code{pi_bar}: A numeric matrix. A \eqn{T \times K} matrix of posterior
change-point location probabilities.
\item \code{v_bar}: A numeric matrix. A \eqn{T \times K} matrix of posterior
rate parameters.
\item \code{u_bar}: A numeric vector. A length \eqn{T} vector of posterior shape
parameters.
\item \code{lambda_bar}: A numeric matrix. A \eqn{T \times K} matrix of
posterior precision signals \eqn{E[\lambda_{kt}|\mathbf{y}_{1:T}]}.
}
}
}
\description{
Fits the univariate version of the MICH model for mean and variance
change-points. Number of change-points can either be fixed or \code{mich_vector()}
will search for the number of changes that maximizes the ELBO when
\code{(L_auto | K_auto | J_auto) == TRUE}.
}
