% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mu2_bar_fn}
\alias{mu2_bar_fn}
\title{Expected Squared-Mean Signal}
\usage{
mu2_bar_fn(b, omega, prob)
}
\arguments{
\item{b}{A numeric vector. Length \eqn{T} vector of conditional mean
parameters.}

\item{omega}{A numeric vector. Length \eqn{T} vector of conditional variance
parameters.}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric vector. A length \eqn{T} vector of \eqn{E[\mu_t]}.
}
\description{
Given that \eqn{E[y_t|b,\tau]=\mu_t=bI(t\geq \tau)}, \eqn{P(\tau=t)=\pi_t},
\eqn{\bar{b}_t=E[b|\tau=t]}, and \eqn{\bar{\omega}_t=V(b|\tau=t)},
\code{mu2_bar_fn()} calculates \eqn{E[\mu^2_t]} as
\eqn{\Sigma_{t'=1}^t (\bar{b}^2_{t'} + 1 / \bar{\omega}_{t'})\pi_{t'}}.
}
\keyword{internal}
