% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mu_bar_fn}
\alias{mu_bar_fn}
\title{Expected Mean Signal}
\usage{
mu_bar_fn(b, prob)
}
\arguments{
\item{b}{A numeric vector. Length \eqn{T} vector of conditional mean
parameters.}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric vector. A length \eqn{T} vector of \eqn{E[\mu_t]}.
}
\description{
Given that \eqn{E[y_t|b,\tau]=\mu_t=bI(t\geq \tau)}, \eqn{P(\tau=t)=\pi_t},
and \eqn{\bar{b}_t=E[b|\tau=t]}, \code{mu_bar_fn()} calculates \eqn{E[\mu_t]} as
\eqn{\Sigma_{t'=1}^t \bar{b}_{t'}\pi_{t'}}.
}
\keyword{internal}
