% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mu_lambda_fn}
\alias{mu_lambda_fn}
\title{Expected Mean-Precision Signal}
\usage{
mu_lambda_fn(b, u, v, prob)
}
\arguments{
\item{b}{A numeric vector. Length \eqn{T} vector of conditional mean
parameters.}

\item{u}{A numeric vector. Length \eqn{T} vector of conditional shape
parameters}

\item{v}{A numeric vector. Length \eqn{T} vector of conditional rate
parameters}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric vector. A length \eqn{T} vector of
\eqn{E[\mu_t\lambda_t]}.
}
\description{
Given that \eqn{E[y_t|b,\tau]=\mu_t=bI(t\geq \tau)},
\eqn{\bar{b}_t=E[b|\tau=t]}, \eqn{\bar{\omega}_t=V(b|\tau=t)},
\eqn{V(y_t|s,\tau)=1/\lambda_t=1/s^{I(t\geq \tau)}},
\eqn{\bar{u}_t/\bar{v}_t=E[s|\tau=t]}, and \eqn{P(\tau=t)=\pi_t},
\code{mu_lambda_fn()} calculates \eqn{E[\mu_t\lambda_t]} as
\eqn{\Sigma_{t'=1}^t \bar{b}_{t'}\bar{v}_{t'} \pi_{t'}/ \bar{u}_{t'}}.
}
\keyword{internal}
