% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_mean_scp}
\alias{multi_mean_scp}
\title{Mulitvariate Mean Single Change-Point Model}
\usage{
multi_mean_scp(y, omega_bar, log_omega_bar, log_pi)
}
\arguments{
\item{y}{A numeric matrix. \eqn{T\times d} matrix of observations with a
single mean change.}

\item{omega_bar}{A numeric vector. Posterior precision parameters
\eqn{\bar{\omega}_t} such that
\eqn{\bar{\Omega}_t = \bar{\omega}_t\mathbf{I}_d}.}

\item{log_omega_bar}{A numeric vector. Log of \code{omega_bar}.}

\item{log_pi}{A numeric vector. Vector of log prior probabilities for the
location of the change-point.}
}
\value{
A list. A list of posterior parameters including the mean \code{b_bar},
and posterior probabilities of the change-point location \code{pi_bar}.
}
\description{
Implementation of the multivariate Mean-SCP model from Berlind, Cappello,
and Madrid Padilla (2025). The function \code{mean_scp()} takes a \eqn{T\times d}
matrix \eqn{\mathbf{y}_{1:T}} with a single mean change and returns the
posterior distribution of the change-point.
}
