% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_mich_cpp}
\alias{multi_mich_cpp}
\title{Multivariate MICH Algorithm}
\usage{
multi_mich_cpp(
  y,
  mu_0,
  fit_intercept,
  refit,
  max_iter,
  tol,
  verbose,
  omega_l,
  log_pi_l,
  omega_bar_l,
  log_omega_bar_l,
  post_params
)
}
\arguments{
\item{y}{A numeric matrix. \eqn{T \times d} matrix of observations.}

\item{mu_0}{A numeric vector. Vector of intercept parameters.}

\item{fit_intercept}{A logical. If \code{fit_intercept == TRUE}, then an
intercept is estimated and \code{mu_0} gets updated.}

\item{refit}{A logical. If \code{refit == TRUE}, then the MICH algorithm is
initialized by the fit provided in \code{post_params}, otherwise the null
model \eqn{\boldsymbol{\mu}_t = \mathbf{0}} and
\eqn{\bar{\pi}_{\ell t} = 1/T} is used as the initialization.}

\item{max_iter}{An integer. Maximum number of iterations. If ELBO does not
converge before \code{max_iter} is reached, then \code{converged == FALSE} in the
returned fit object.}

\item{tol}{A scalar. Convergence tolerance for relative increase in ELBO.}

\item{verbose}{A logical. If \code{verbose == TRUE}, then the value of the ELBO
is printed every 5000th iteration.}

\item{omega_l}{A scalar. Prior precision parameter for mean-scp components
of model.}

\item{log_pi_l}{A numeric matrix. A \eqn{T \times L} matrix of prior log
change-point location probabilities for each of the \eqn{L} mean
change-points.}

\item{omega_bar_l, log_omega_bar_l}{Numeric vectors. Vector of posterior
precision parameters \eqn{\{\bar{\omega}_t\}_{t=1}^T} and log evaluations
such that \eqn{V(\mathbf{b}_\ell|\tau=t) = \bar{\omega}_t\mathbf{I}_d}.}

\item{post_params}{A list. A length \eqn{L} list of the posterior parameters
of each mean-scp component. Each element of the list is a list containing
a \eqn{T\times L} matrix of posterior mean parameters and length \eqn{T}
vectors of posterior change-point location probabilities and their log
evaluations.}
}
\value{
A List. Parameters of the variational approximation the MICH
posterior distribution, including:
\itemize{
\item \code{L}: An integer. Number of components included in model.
\item \code{residual}: A numeric matrix. Residual \eqn{\mathbf{r}_{1:T}} after
subtracting out each \eqn{E[\boldsymbol{\mu}_{\ell t}]} from
\eqn{\mathbf{y}_{1:T}}.
\item \code{mu}: A numeric matrix. Posterior estimate of
\eqn{\Sigma_{\ell=1}^L E[\boldsymbol{\mu}_{\ell,1:T}|\mathbf{y}_{1:T}]}.
\item \code{mu_0}: A numeric vector. Estimate of the intercept.
\item \code{post_params}: A list. List of posterior parameters for each mean-scp
component.
\item \code{elbo}: A numeric vector. Value of the ELBO after each iteration.
\item \code{converged}: A logical. Indicates whether relative increase in the ELBO
is less than \code{tol}.
}
}
\description{
Implementation of Algorithm 3 from Berlind, Cappello, and Madrid Padilla
(2025). This algorithm takes a sequence of \eqn{d}-dimensional observations
\eqn{\mathbf{y}_{1:T}}, and iteratively fits \eqn{L} mean-scp models
resulting in a variational approximation to the posterior distribution of
the \eqn{L} mean change-points. The algorithm terminates once the
percentage increase in the ELBO falls below \code{tol}.
}
\keyword{internal}
