% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_mu2_bar_fn}
\alias{multi_mu2_bar_fn}
\title{Multivariate Expected Squared-Mean Signal}
\usage{
multi_mu2_bar_fn(b, omega, prob)
}
\arguments{
\item{b}{A numeric matrix. \eqn{T\times d} matrix of conditional mean
parameters.}

\item{omega}{A numeric vector. Length \eqn{T} vector of conditional variance
parameters.}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric matrix. A \eqn{T\times d} matrix of
\eqn{E[b^2_iI(t\geq \tau)]}.
}
\description{
Given that \eqn{E[\mathbf{y}_t|\mathbf{b},\tau] = \mathbf{b}I(t\geq \tau)},
\eqn{P(\tau = t)=\pi_t}, and for some \eqn{1\leq i \leq d},
\eqn{\bar{b}_{it}=E[b_i|\tau=t]} and \eqn{\bar{\omega}_{it}=V(b_i|\tau=t)},
\code{multi_mu2_bar_fn()} calculates \eqn{E[b_i^2I(t\geq \tau)]} as
\eqn{\Sigma_{t'=1}^t (\bar{b}^2_{it'} + 1/\bar{\omega}_{it'}) \pi_{t'}}.
}
\keyword{internal}
