% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_mu_bar_fn}
\alias{multi_mu_bar_fn}
\title{Multivariate Expected Mean Signal}
\usage{
multi_mu_bar_fn(b, prob)
}
\arguments{
\item{b}{A numeric matrix. \eqn{T\times d} matrix of conditional mean
parameters.}

\item{prob}{A numeric vector. Vector of change-point location probabilities.}
}
\value{
A numeric matrix. A \eqn{T\times d} matrix of
\eqn{E[\mathbf{b}I(t\geq \tau)]}.
}
\description{
Given that \eqn{E[\mathbf{y}_t|\mathbf{b},\tau] = \mathbf{b}I(t\geq \tau)},
\eqn{P(\tau = t)=\pi_t}, and \eqn{\bar{\mathbf{b}}_t=E[\mathbf{b}|\tau=t]},
\code{multi_mu_bar_fn()} calculates \eqn{E[\mathbf{b}I(t\geq \tau)]} as
\eqn{\Sigma_{t'=1}^t \bar{\mathbf{b}}_{t'}\pi_{t'}}.
}
\keyword{internal}
