% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prob_check}
\alias{prob_check}
\title{Probability vector check.}
\usage{
prob_check(probs, n, T)
}
\arguments{
\item{probs}{A character, numeric vector, or numeric matrix. If character,
\code{probs} should be equal to 'weighted' or 'uniform'. If vector, elements of
\code{probs} must sum to one. If matrix, rows of \code{probs} must sum to one.}

\item{n}{An integer. Number of columns if \code{probs} is a matrix.}

\item{T}{An integer. Length or number of rows in \code{probs}.}
}
\value{
A character, numeric vector, or numeric matrix. If no error is thrown
during evaluation, then \code{prob_check()} returns the \code{probs} argument.
}
\description{
\code{prob_check()} throws an error if \code{probs} is not either one of the string
arguments 'uniform' or 'weighted', a length \eqn{T} vector with elements that
sum to one, or an \eqn{T \times n} matrix with columns that sum to one.
}
