% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{revcumsum}
\alias{revcumsum}
\title{Reverse cumulative sum}
\usage{
revcumsum(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector. Reverse cumulative sum of the elements of \code{x}.
}
\description{
\code{revcumsum()} takes a length \eqn{n} vector \eqn{\{x_i\}_{i=1}^n} and returns \eqn{\{\Sigma_{j=i}^n x_j\}_{i=1}^n}.
}
\examples{
revcumsum(1:10)
}
