% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scalar_check}
\alias{scalar_check}
\title{Scalar value check.}
\usage{
scalar_check(x)
}
\arguments{
\item{x}{A scalar.}
}
\value{
A scalar If no error is thrown during evaluation, then
\code{scalar_check()} returns the \code{x} argument.
}
\description{
\code{scalar_check()} throws an error if \code{x} is not equal to a scalar value.
}
