% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mich.fit.R
\name{summary.mich.fit}
\alias{summary.mich.fit}
\title{Summary method for mich.fit objects}
\usage{
\method{summary}{mich.fit}(object, level = 0.95, max_length = NULL, ...)
}
\arguments{
\item{object}{A mich.fit object. Output of running \code{mich()} on a numeric vector
or matrix.}

\item{level}{A scalar. A number between (0,1) indicating the significance
level to construct credible sets at.}

\item{max_length}{An integer. Detection threshold, see \code{mich_sets()}. Equal
to \code{log(T)^2} by default.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
A list. A list of summary quantities including:
\itemize{
\item \code{elbo}: The value of the ELBO for the model.
\item \code{converged}: Indicator for whether the model has converged.
\item \code{level}: The significance level used to construct credible sets.
\item \code{L},\code{K},\code{J}: The number of mean-scp, var-scp, and meanvar-scp components
included in the model.
\item \code{mean_cp},\code{var_cp},\code{meanvar_cp}: Lists with \code{cp}, the estimated
change-points, and \code{sets}, their corresponding \code{level}-level credible
sets.
}
}
\description{
Prints a summary of the resulting fit from calling \code{mich()}, including the
ELBO for the fitted model, estimated change-points with \code{level}-level
credible sets.
}
