% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{well_log}
\alias{well_log}
\title{Shankle Oil Well Lithology}
\format{
\subsection{\code{well_log}}{

A data frame with 346 rows and 8 columns:
\describe{
\item{Facies}{1: "Nonmarine Sandstone, 2: "Nonmarine Coarse Siltstone, 3: Nonmarine Fine Siltstone, 4: Marine Siltstone/Shale, 5: Mudstone, 6: Wackestone, 7: Dolomite, 8: Packstone-Grainstone,8: Packstone-Grainstone"}
\item{Depth}{Well depth in feet}
\item{GR}{Gamma ray emissions}
\item{Rt}{Deep resistivity}
\item{DeltaPhi}{Difference of neutron and density porosity}
\item{AvgPhi}{Average neutron and density porosity}
\item{PE}{Photoelectric factor}
\item{MnM}{Marine/non-marine indicator}
}
}
}
\source{
Bohling, G. and M. Dubois (2003). An integrated application of neural
network and markov chain techniques to the prediction of lithofacies from
well logs: Kansas geological survey open-file report 2003-50, 6 p. Group 6.
}
\usage{
well_log
}
\description{
Six petrophysical measures collected by the Kansas Geological Survey by
drilling into the B1-C formations of the Shankle well in the Panoma oil and
gas field of southwest Kansas.
}
\keyword{datasets}
