% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{log_mean_prior}
\alias{log_mean_prior}
\title{Log Weighted Mean-SCP Prior}
\usage{
log_mean_prior(T, d)
}
\arguments{
\item{T}{An integer. Number of observations (rows) of \eqn{\mathbf{y}_{1:T}}.}

\item{d}{An integer. Dimension (columns) of \eqn{\mathbf{y}_{1:T}}.}
}
\value{
A numeric vector. Log prior probabilities.
}
\description{
Log weighted prior for the Mean-SCP model as described in Appendix C.2 of
Berlind, Cappello, and Madrid Padilla (2025). This prior ensures that in the
absence of a change-point, the posterior probabilities are approximately
uniform.
}
