% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mich.R
\name{mich}
\alias{mich}
\title{Title}
\usage{
mich(
  y,
  fit_intercept = TRUE,
  fit_scale = TRUE,
  standardize = TRUE,
  J = 0,
  L = 0,
  K = 0,
  J_auto = FALSE,
  L_auto = FALSE,
  K_auto = FALSE,
  J_max = Inf,
  L_max = Inf,
  K_max = Inf,
  tol = 1e-05,
  merge_prob = NULL,
  merge_level = 0.95,
  max_iter = 10000,
  verbose = FALSE,
  reverse = FALSE,
  restart = TRUE,
  increment = 1,
  omega_j = 0.001,
  u_j = 0.001,
  v_j = 0.001,
  pi_j = "weighted",
  omega_l = 0.001,
  pi_l = "weighted",
  u_k = 0.001,
  v_k = 0.001,
  pi_k = "weighted"
)
}
\arguments{
\item{pi_k}{}
}
\description{
Title
}
