% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mich_sets.R
\name{mich_sets}
\alias{mich_sets}
\title{MICH Posterior Credible Sets}
\usage{
mich_sets(probs, max_length = log(nrow(probs))^1.5, level = 0.9)
}
\arguments{
\item{probs}{A numeric Matrix. A \eqn{T \times N} matrix of posterior
probabilities for the location of the change-points.}

\item{max_length}{A positive scalar. Detection threshold, if a credible set
contains more that \code{max_length} indices, then no change is detected. Set
equal to \eqn{\log^{1.5} T} by default (see Section 2.5 of Berlind,
Cappello, and Madrid Padilla (2025)).}

\item{level}{A scalar. A single number in (0,1) that gives the lower bound
for the probability that each credible set contains a change-point.}
}
\value{
A list. MAP estimator of each change-point and corresponding credible
set.
}
\description{
The function \code{mich_sets()} takes a \eqn{T \times N} matrix of posterior
change-point location probabilities \eqn{\{\pi_{i,1:T}\}_{i=1}^N}, a coverage
level \eqn{\alpha}, and a max set length, and for each \eqn{1\leq i \leq N}
returns the MAP estimator \eqn{\arg\max_{1\leq t \leq T} \pi_{it}} and the
smallest set \eqn{S_i} such that \eqn{\Sigma_{t \in S_i} \pi_{it} \geq \alpha}
if \eqn{|S_i|<} \code{max_length}.
}
