% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mich.fit.R
\name{plot.mich.fit}
\alias{plot.mich.fit}
\title{Plot method for mich.fit objects}
\usage{
\method{plot}{mich.fit}(
  fit,
  level = 0.95,
  max_length = NULL,
  signal = FALSE,
  cs = TRUE,
  n_plots = 5
)
}
\arguments{
\item{fit}{A mich.fit object. Output of running \code{mich()} on a numeric vector
or matrix.}

\item{level}{A scalar. A number between (0,1) indicating the significance
level to construct credible sets at when \code{cs == TRUE}.}

\item{max_length}{An integer. Detection threshold, see \code{mich_sets()}. Equal
to \eqn{\log(T)^{1.5}} by default.}

\item{signal}{A logical. If \code{TRUE}, then the posterior mean and precision
signals are also plotted.}

\item{cs}{A logical. If \code{TRUE}, then \code{level}-level credible sets for each
detected change-point are also plotted.}

\item{n_plots}{An integer. Number of plot to display at once when data \code{y} is
a matrix.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plot method for mich.fit objects
}
\examples{
y = c(rnorm(100,0,10), rnorm(100,10,3), rnorm(100,0,6))
fit = mich(y, J = 2) # fit two mean-variance change-points
plot(fit, level = 0.95, cs = TRUE) # plot change-points with 95\% credible sets
fit = mich(y, J = 1, L = 1) # fit one mean and one mean-variance change-point
plot(fit, level = 0.95, cs = TRUE, signal = TRUE) # plot change-points with 95\% credible sets and signal

}
